webDriver Solver
